imgens de cartoes e finalização da area de pedidos locais

+++ app/Http/Controllers/RequestsController.php
@@ -6,6 +6,7 @@ use App\Classes\Tools;
 use App\Events\notificationNewRequest;
 use App\Models\AdditionalItemModel;
 use App\Models\ItemModel;
+use App\Models\PaymentMethodsModel;
 use App\Models\RequestAdditionalItemModal;
 use App\Models\RequestsItemsModel;
 use App\Models\RequestsModel;
@@ -24,7 +25,7 @@ class RequestsController extends Controller
     public function add_item_request(Request $request)
     {
         $data = $request->all();
-        $request = RequestsModel::select('id')->where('table', $data['table'])->where('client_name', $data['client'])->first();
+        $request = RequestsModel::select('id')->where('table', $data['table'])->where('client_name', $data['client'])->where('status', 1)->first();
         $product = ItemModel::select('value')->find($this->Tools->hash($data['item'], 'decrypt'));
 
         for ($i = 1; $i <= $data['amount']; $i++) {
@@ -44,7 +45,7 @@ class RequestsController extends Controller
     public function send_item_request(Request $request)
     {
         $data = $request->all();
-        $request = RequestsModel::select('id')->where('table', $data['table'])->where('client_name', strtoupper($data['client']))->first();
+        $request = RequestsModel::select('id')->where('table', $data['table'])->where('client_name', strtoupper($data['client']))->where('status', 1)->first();
 
         if (RequestsItemsModel::where('request_id', $request->id)->where('status', 1)->update(['waiter' => session('user')['name'], 'status' => 2])) {
             event(new notificationNewRequest([
@@ -105,18 +106,21 @@ class RequestsController extends Controller
     public function save_obs_item_request(Request $request)
     {
         $inputs = $request->all();
-        foreach ($inputs['additionals'] as $input) {
-            if ($input['check'] == 'true') {
-                $additional = AdditionalItemModel::select('value')->find($input['id']);
-                $additionals = [
-                    'additional_id' => $input['id'],
-                    'item_id' => $this->Tools->hash($inputs['id'], 'decrypt'),
-                    'value' => $additional->value,
-                ];
-                RequestAdditionalItemModal::updateOrCreate(['additional_id' => $input['id'], 'item_id' => $this->Tools->hash($inputs['id'], 'decrypt')], $additionals);
-            } else {
-                RequestAdditionalItemModal::where('additional_id', $input['id'])->where('item_id', $this->Tools->hash($inputs['id'], 'decrypt'))->delete();
+        if (isset($inputs['additionals'])) {
+            foreach ($inputs['additionals'] as $input) {
+                if ($input['check'] == 'true') {
+                    $additional = AdditionalItemModel::select('value')->find($input['id']);
+                    $additionals = [
+                        'additional_id' => $input['id'],
+                        'item_id' => $this->Tools->hash($inputs['id'], 'decrypt'),
+                        'value' => $additional->value,
+                    ];
+                    RequestAdditionalItemModal::updateOrCreate(['additional_id' => $input['id'], 'item_id' => $this->Tools->hash($inputs['id'], 'decrypt')], $additionals);
+                } else {
+                    RequestAdditionalItemModal::where('additional_id', $input['id'])->where('item_id', $this->Tools->hash($inputs['id'], 'decrypt'))->delete();
+                }
             }
+
         }
         $observation = RequestsItemsModel::find($this->Tools->hash($inputs['id'], 'decrypt'));
         $observation->observation = $inputs['obs'];
@@ -239,7 +243,6 @@ class RequestsController extends Controller
         }
         return $clients;
     }
-
     public function requests_client_view(Request $request)
     {
         $id = $request->get('id');
@@ -253,6 +256,117 @@ class RequestsController extends Controller
             'pending' => $requestItems,
         ];
         return $data;
+    }
+    public function request_finish($id)
+    {
+        if (RequestsModel::find($this->Tools->hash($id, 'decrypt'))) {
+            return 'true';
+        }
+    }
+    public function view_item_request(Request $request)
+    {
+        return RequestsItemsModel::find($this->Tools->hash($request->get('id'), 'decrypt'));
+    }
+// PAGAMENTO
+    public function finalize_payment(Request $request)
+    {
+        $data = $request->all();
+
+        // ENVIA DADOS PARA A RECEITA
+        // RETORNA SE DEU CERTO
+
+        // FINALIZA ITENS NO SISTEMA
+        //  PAGAMENTO DIVIDIDO
+        if ($data['split_payment']['active'] == 'true') {
+            RequestsItemsModel::whereIn('id', $data['split_payment']['items'])->update(['status' => 4, 'payment_method' => $data['method']]);
+            if (RequestsItemsModel::where('request_id', $this->Tools->hash($data['id'], 'decrypt'))->where('status', 3)->count() < 1) {
+                RequestsModel::find($this->Tools->hash($data['id'], 'decrypt'))->update(['status' => 2]);
+                RequestsItemsModel::where('request_id', $this->Tools->hash($data['id'], 'decrypt'))->where('status', 2)->delete();
+            }
+
+            return 'split_success';
+        } else {
+            if (RequestsItemsModel::where('request_id', $this->Tools->hash($data['id'], 'decrypt'))->where('status', 3)->update(['status' => 4, 'payment_method' => $data['method']])) {
+                RequestsModel::find($this->Tools->hash($data['id'], 'decrypt'))->update(['status' => 2]);
+                RequestsItemsModel::where('request_id', $this->Tools->hash($data['id'], 'decrypt'))->where('status', 2)->delete();
+            }
+            return 'success';
+
+        }
+    }
+    public function tax_coupon(Request $request)
+    {
+        $data = $request->all();
+
+        if ($data['split_payment']['active'] == 'true') {
+            $requests = RequestsItemsModel::with('additionals')->whereIn('id', $data['split_payment']['items'])->where('status', 4)->where('print', null)->orderBy('product_id', 'asc')->get();
+            RequestsItemsModel::whereIn('id', $data['split_payment']['items'])->where('status', 4)->where('print', null)->update(['print' => 1]);
+        } else {
+            $requests = RequestsItemsModel::with('additionals')->where('request_id', $this->Tools->hash($data['id'], 'decrypt'))->where('status', 4)->where('print', null)->orderBy('product_id', 'asc')->get();
+            RequestsItemsModel::where('request_id', $this->Tools->hash($data['id'], 'decrypt'))->where('status', 4)->where('print', null)->update(['print' => 1]);
+        }
+
+        if ($data['action'] == 'not') {
+            return false;
+        }
+
+        $method = PaymentMethodsModel::select('name')->find($data['method']);
+        $items = [];
+        foreach ($requests as $item) {
+            if (isset($total)) {
+                $total += $this->Tools->sum_values_item_number($item->id);
+            } else {
+                $total = $this->Tools->sum_values_item_number($item->id);
+            }
+            if ($item->additionals != '[]' || $item->observation) {
+                $items[] = [
+                    'name' => $item->product->name,
+                    'val_un' => 'R$' . number_format($item->product->value, 2, ',', '.'),
+                    'val_total' => 'R$' . number_format($item->product->value, 2, ',', '.'),
+                    'additionals' => $item->additionals,
+                    'amount' => '1',
+                ];
+
+            } else {
+                if (isset($count[$item->product->id])) {
+                    $count[$item->product->id]++;
+                } else {
+                    $count[$item->product->id] = 1;
+                }
+                if (isset($sum[$item->product->id])) {
+                    $sum[$item->product->id] += $item->product->value;
+                } else {
+                    $sum[$item->product->id] = $item->product->value;
+                }
+
+                $items[$item->product->id . 'item'] = [
+                    'name' => $item->product->name,
+                    'val_un' => 'R$' . number_format($item->product->value, 2, ',', '.'),
+                    'val_total' => 'R$' . number_format($sum[$item->product->id], 2, ',', '.'),
+                    'additionals' => [],
+                    'amount' => $count[$item->product->id],
+                ];
+            }
+        }
+        $response = [
+            'command' => RequestsModel::find($this->Tools->hash($data['id'], 'decrypt')),
+            'items' => $items,
+            'total' => 'R$' . number_format($total, 2, ',', '.'),
+            'method' => $method,
+
+        ];
+
+        switch ($data['action']) {
+            case 'email':
+                # code...
+                break;
+            case 'whatsapp':
+                # code...
+                break;
+            case 'print':
+                return view('app.component.non-tax-coupon', $response);
+                break;
+        }
     }
 // TABELAS
     public function request_client_table(Request $request)
@@ -300,6 +414,47 @@ class RequestsController extends Controller
             $dado[] = $this->Tools->sum_values_item($item->id);
             $dado[] = '<button onclick="return additional_item_request(\'' . $this->Tools->hash($item->product_id, 'encrypt') . '\',\'' . $this->Tools->hash($item->id, 'encrypt') . '\')" class="btn btn-sm btn-primary" ><i class="fa-solid fa-pen"></i></button> <button onclick="return  delete_item_request(\'' . $this->Tools->hash($item->id, 'encrypt') . '\')" class="btn btn-sm btn-danger m-t-3"><i class="fa-solid fa-trash"></i></button>';
             // $dado[] = $item->description;'R$' . number_format($item->value, 2, ',', '.')
+            $dados[] = $dado;
+        }
+
+        //Cria o array de informações a serem retornadas para o Javascript
+        $json_data = array(
+            "draw" => intval($requestData['draw']), //para cada requisição é enviado um número como parâmetro
+            "recordsTotal" => intval($filtered), //Quantidade de registros que há no banco de dados
+            "recordsFiltered" => intval($rows), //Total de registros quando houver pesquisa
+            "data" => $dados, //Array de dados completo dos dados retornados da tabela
+        );
+
+        return json_encode($json_data); //enviar dados como formato json
+    }
+    public function split_payment_table(Request $request)
+    {
+        $requestData = $request->all();
+        $columns = array(
+            0 => 'id',
+            1 => 'product_id',
+            2 => 'value',
+            3 => 'id',
+        );
+
+        if ($requestData['columns'][1]['search']['value']) {
+            $items = RequestsItemsModel::with('product')->where('request_id', $this->Tools->hash($requestData['columns'][1]['search']['value'], 'decrypt'))->where('status', 3)->orderBy($columns[$requestData['order'][0]['column']], $requestData['order'][0]['dir'])->get();
+        } else {
+            $items = array();
+        }
+        $rows = 0;
+        $filtered = count($items);
+        $dados = array();
+
+        foreach ($items as $item) {
+            $dado = array();
+            $dado[] = '#' . $item->id;
+            $dado[] = $item->product->name;
+            $dado[] = $this->Tools->sum_values_item($item->id);
+            $dado[] = '<div class="custom-control custom-checkbox">
+                       <input class="custom-control-input custom-control-input-secondary" type="checkbox" name="item" id="item' . $item->id . '" value="' . $item->id . '">
+                       <label for="item' . $item->id . '" class="custom-control-label"></label>
+                       </div>';
             $dados[] = $dado;
         }
 
@@ -373,7 +528,7 @@ class RequestsController extends Controller
         if ($id) {
             $items = RequestsItemsModel::with('product')->select('product_id', DB::raw('COUNT(id) as count'))
                 ->where('request_id', $this->Tools->hash($id, 'decrypt'))
-                ->whereBetween('status', [2, 3])
+                ->where('status', 3)
                 ->groupBy('product_id')
                 ->orderBy('count', $requestData['order'][0]['dir'])
                 ->get();
@@ -551,7 +706,7 @@ class RequestsController extends Controller
             $dado[] = $item->product->name;
             $dado[] = $item->waiter;
             $dado[] = $this->Tools->sum_values_item($item->id);
-            $dado[] = $item->status != 2 ? '<button onclick="return  delete_item_request(\'' . $this->Tools->hash($item->id, 'encrypt') . '\')" class="btn btn-sm btn-danger m-t-3"><i class="fa-solid fa-trash"></i></button>' : '<button onclick="return additional_item_request(\'' . $this->Tools->hash($item->product_id, 'encrypt') . '\',\'' . $this->Tools->hash($item->id, 'encrypt') . '\')" class="btn btn-sm btn-primary" ><i class="fa-solid fa-pen"></i></button> <button onclick="return  delete_item_request(\'' . $this->Tools->hash($item->id, 'encrypt') . '\')" class="btn btn-sm btn-danger m-t-3"><i class="fa-solid fa-trash"></i></button>';
+            $dado[] = $item->status != 2 ? '<button onclick="return  delete_item_request(\'' . $this->Tools->hash($item->id, 'encrypt') . '\')" class="btn btn-sm btn-secondary m-t-3"><i class="fa-solid fa-eye"></i></button> <button onclick="return view_item_request(\'' . $this->Tools->hash($item->id, 'encrypt') . '\')" class="btn btn-sm btn-danger m-t-3"><i class="fa-solid fa-trash"></i></button>' : '<button onclick="return additional_item_request(\'' . $this->Tools->hash($item->product_id, 'encrypt') . '\',\'' . $this->Tools->hash($item->id, 'encrypt') . '\')" class="btn btn-sm btn-primary" ><i class="fa-solid fa-pen"></i></button> <button onclick="return  delete_item_request(\'' . $this->Tools->hash($item->id, 'encrypt') . '\')" class="btn btn-sm btn-danger m-t-3"><i class="fa-solid fa-trash"></i></button>';
             $dados[] = $dado;
         }
